Update Musa subtaxa
The subtaxa list contains 18 entries that do not capture de diversity of the genus and many cultivated bananas used for food and agriculture and some of their crop wild relatives.

It looks like the script only loop on id 4641 and its subtaxa
But the Musa genus is much larger

There are at least 45 taxa that are not captured at the moment. A bit more if taking all subtaxa automatically under Musa.